Update time of meeting
As discussed in the agenda and in person we actually stick to PST for meetings, meaning in other places due to day light saving and the the like, the time changes. Swapped out bi-weekly as in various places this can mean _either_ [once every two weeks, or twice a week](https://en.wiktionary.org/wiki/biweekly). The English language is rubbish sometimes.